Solve Part 1 of Day 18
So this only works on a technicality of the input. If it weren't for
the fact that all of the input coordinates were two digits at best, this
would run into a bug of having two cubes get tagged as adjacent
even if they really weren't. But it works, so who cares?